+++ scripts/Opencast/setup_opencast.sh
@@ -19,10 +19,14 @@
 ######################################################################################
 #
 # Note, that this script must be executed as superuser with:
-#   sudo sh setup_opencast.sh <java_version> <opencast_version>
+#   sudo sh setup_opencast.sh <java_version> <opencast_version> <force_yes>
 #
 # java_version:     The java version to install (e.g., 11). Note, that for Opencast 10 and newer JDK 11 is supported only (state on 28.07.22).
 # opencast_version: The stable Opencast major version to install (e.g., 11 (see https://opencast.org/category/releases/)).
+# force_yes:        Optional parameter, that can be passed.
+#                   If force_yes is passed for this parameter, the script will be executed without
+#                   a confirmation by the user.
+#                   If another value is passed for this parameter, this parameter is ignored.
 #
 # This script prepares the system for this Opencast installation and installs this Opencast version afterwards.
 # See https://docs.opencast.org/r/11.x/admin/#installation/debs/ for further details.
@@ -35,18 +39,23 @@ echo ""
 setupdir="$(dirname "${0}")"
 java_version="${1}"
 opencast_version="${2}"
+force_yes="${3:-ignore}"
 
 echo "The stable Opencast major version ${opencast_version} will be installed."
 echo ""
-read -p "Enter 'yes' to proceed or another text to abort:" userdecision
-echo ""
 
-if [ -z $userdecision ] || [ $userdecision != "yes" ]
+if [ $force_yes != "force_yes" ]
 then
-    echo "+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"
-    echo " ==> Aborted install of Opencast."
-    echo "+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"
-    exit 0
+  read -p "Enter 'yes' to proceed or another text to abort:" userdecision
+  echo ""
+
+  if [ -z $userdecision ] || [ $userdecision != "yes" ]
+  then
+      echo "+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"
+      echo " ==> Aborted install of Opencast."
+      echo "+++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++"
+      exit 0
+  fi
 fi
 
 sh "${setupdir}/prepare_opencast_install.sh" "${java_version}" "${opencast_version}"
